Moved `DynamicEvent`, and `DynamicEventSender`, and supporting types to their own files
